Refactor: Generalized add to cart functionality
